requirements.txt
# Pinterest Standalone Scraper Requirements
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
selenium>=4.10.0
undetected-chromedriver>=3.5.0
requests>=2.28.0
Pillow>=9.0.0
pymongo>=4.0.0
google-generativeai>=0.3.0
loguru>=0.7.0
